no errors, hardcoded values
