+++ src/components/Funds.jsx
@@ -1,0 +1,14 @@
+import React from "react"
+
+
+function Funds() {
+
+  return (
+    <div>
+      <h3>Funds</h3>
+      <p>5000£</p>
+    </div>
+  )
+}
+
+export default Funds
