--- src/components/WalletPage/style.css
@@ -1,23 +1,0 @@
-.container {
-  width: 1000px;
-  height: 300px;
-  background-color: aquamarine;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.budget-box {
-  width: 200px;
-  height: 200px;
-  background-color: antiquewhite;
-
-}
-
-
-
-.purchasedList{
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
