--- src/pages/Wallet.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Typography } from '@mui/material';
-
-const Wallet = () => {
-  return (
-    <Typography variant="h6" component="div">
-      This is Wallet page
-    </Typography>
-  );
-};
-
-export default Wallet;
